save SkillTrainingSession data in SkillProgressRecord

+++ src/main/java/co/uk/bransby/equinetrainingtrackerapi/api/models/SkillTrainingSession.java
@@ -39,6 +39,6 @@ public class SkillTrainingSession {
 
     private ProgressCode progressCode;
 
-    private Long trainingTime;
+    private Integer trainingTime;
     private String notes;
 }
